Variation & Training Support
This branch allows for multiple variations to be completed instead of one. Users can vote for their favourite variation and train their datasets. This changes ProntoLM from a simple LM to an ML as it's now human trainable.

Although this falls under the term ML, I still don't consider it ML yet. I believe in order for this program to truely be ML, it needs the ability to self improve without human training (i.e. access to the internet to improve its dataset).